% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEnricher.R
\name{dEnricher}
\alias{dEnricher}
\title{double enrich}
\usage{
dEnricher(pair1, pair2, background = NULL, rds = NULL, cores = 6,
  refresh = FALSE)
}
\arguments{
\item{pair1}{pair1}

\item{pair2}{pair2}

\item{background}{background set}

\item{rds}{rds file path}

\item{cores}{cpu core numbers}

\item{refresh}{re caculate}
}
\value{
XY2Z
}
\description{
Enrich two sets
}
\author{
Navy Cheng
}
