% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2yMatrixAdjust.R
\name{getX2yMatrixAdjust}
\alias{getX2yMatrixAdjust}
\title{Get xy.struct}
\usage{
getX2yMatrixAdjust(xa, xb, background = NULL)
}
\arguments{
\item{xa}{x2b paris or list}

\item{xb}{x2b paris or list}

\item{background}{a set of `b`}
}
\value{
\item{name fix.list}{description a list contain the fixed result}
}
\description{
Get xy.struct
}
\author{
c509
}
