% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openfoodfacts}
\alias{openfoodfacts}
\title{Grocery products facts.}
\format{
A data frame with 1000 rows and 59 variables:
\describe{
\item{product_name}{product name}
\item{nutriscore_grade}{a grading of the nutricial quality of the product}
\item{nova_group}{a grading of the additive present in the product}
\item{*_100g}{quantity of element per 100g of the product}
...
}
}
\source{
\url{https://world.openfoodfacts.org/data/data-fields.txt}
}
\usage{
openfoodfacts
}
\description{
A dataset containing the ingredients and other attributes of a sample of 200
grocery store products
}
\keyword{datasets}
