% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.R
\name{step_mutual_information}
\alias{step_mutual_information}
\title{reciepe step for removing low mutual information columns form a dataframe}
\usage{
step_mutual_information(df, mutual_information = 80, ...)
}
\arguments{
\item{df}{dataframe}

\item{mutual_information}{mutual information}

\item{...}{other parameters passed onto dx2y function}
}
\value{

}
\description{
reciepe step for removing low mutual information columns form a dataframe
}
