% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2yMatrix.R
\name{x2yMatrix}
\alias{x2yMatrix}
\title{Change x2y struct to xy.matrix}
\usage{
x2yMatrix(a2b)
}
\arguments{
\item{a2b}{Input data with a x2y struct}
}
\value{
\item{name xy.matrix}{description a value transformed from x2y}
}
\description{
Change x2y struct to xy.matrix
}
\details{
x2y struct is a list like this

\code{
List of 3:\cr
  $ x1: chr[1:3] 'y1' 'y2' 'y3'\cr
  $ x2: chr[1:2] 'y2' 'y3'\cr
  $ x3: chr[1:2] 'y1' 'y2'\cr
}

xy.matrix struct is a matrix like this

\code{
    x1  x2  x3\cr
y1  1   0   1\cr
y2  1   1   1\cr
y3  1   1   0\cr
}
}
\examples{
a2b <- list(x1=c('y1','y2','y3'), x2=c('y2','y3'), x3=c('y1','y2'))
x2yMatirx(a2b)
}
\keyword{filename}
