% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyCor.R
\name{xyCor}
\alias{xyCor}
\title{Calculate the relationship of tow x2y struct}
\usage{
xyCor(am, bm, verbose = F)
}
\arguments{
\item{am}{xa, input data with a fixed x2y struct}

\item{bm}{xb, input data with a fixed x2y struct}
}
\value{
\item{name xy.relationship}{Phi, pvalue and contingency table of xa and xb}
}
\description{
Calculate the relationship of tow x2y struct
}
\details{
The tow x2y struct contain the same 'y' set {Y}. The contingency table represent the relationship of 'x' pairs with 'y'
1 represent xa or xb have associate with 'y', and 0 represent no association. So, 11 represent both xa and xb have
associate with 'y' and 0 represent neither of them have association.

Phi and p-value of xa and xb is also caculated.

\code{
          b\cr
          1   0\cr           
   a   1  11  10\cr
       0  1   0\cr
}
}
\examples{
xa2y <- list(x1=c('y1','y2','y3'), x2=c('y2','y3'), x3=c('y1','y2'))
xa2b.m <- x2yMatirx(a2b)
xb2y <- list(x1=c('y1','y2','y3','y4'), x2=c('y2','y3'), x3=c('y1','y2','y5'))
xb2b.m <- x2yMatirx(a2b)
fix <- x2yMatrixAdjust(xa2b.m, xb2b.m)
xyCor(fix$a, fix$b)
}
