% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x2y.R
\name{dx2y}
\alias{dx2y}
\title{Compute the mutual information between dataframe columns}
\usage{
dx2y(d, target = NA, confidence = FALSE, sample_n = 5000)
}
\arguments{
\item{d}{dataframe to compare columns}

\item{target}{(optional) the outcome column if any}

\item{confidence}{Boolean, shall we compute the 95\% Confidence interval (default FALSE)}

\item{sample_n}{max number of sample to consider (default 50k )}
}
\value{
a dataframe with columns \code{x} and \code{y} with all combinations of column pairs
and for each pair a \code{perc_of_obs} for the percentage of observations present,
and  \code{x2y} value of the mutual information between those two columns
}
\description{
Compute the mutual information between dataframe columns
}
\examples{
pairwise_iris <- dx2y(iris)
head(pairwise_iris)
}
